backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy<2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4
python-multipart==0.0.6
pydantic<2
python-magic==0.4.27
python-magic-bin==0.4.14
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
